+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 * 【db     】      修复JndiDSFactory空指针问题
 * 【core   】      修复BiMap.put错误的返回值（pr#1218@Gitee）
 * 【core   】      修复BooleanUtil.andOfWrap针对null错误问题（issue#3587@Github）
+* 【core   】      修复FileUtil#getTotalLines在JDK9+结果错误问题（issue#3591@Github）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.8.27(2024-03-29)